Add UI for multi root users
