+++ ui/src/Pages/SettingsPage/Users/UsersTable.js
@@ -4,7 +4,7 @@ import { Table, Loader, Button, Icon } from 'semantic-ui-react'
 import { useQuery, gql } from '@apollo/client'
 import UserRow from './UserRow'
 import AddUserRow from './AddUserRow'
-import { SectionTitle } from './SettingsPage'
+import { SectionTitle } from '../SettingsPage'
 
 const USERS_QUERY = gql`
   query settingsUsersQuery {
@@ -13,7 +13,7 @@ const USERS_QUERY = gql`
       username
       # rootPath
       admin
-      albums {
+      rootAlbums {
         id
         filePath
       }
